requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
IMDbPY==2021.4.18
itsdangerous==2.1.1
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.1
pycparser==2.21
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.9
Werkzeug==2.0.3
WTForms==3.0.1
